% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{pl_blank_session}
\alias{pl_blank_session}
\title{pl_blank_session}
\usage{
pl_blank_session(pl_labs = NULL)
}
\arguments{
\item{pl_labs}{Optional list of ggplot plotting labels}
}
\value{
A ggplot plotting object with a datetime x-axis from over the max range of sessions and a discrete y-axis over the range of work focus codes.
}
\description{
#' Uses a generic title. Add subtitle, y-axis label, etc through `pl_labs`
}
