% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-utils.R
\name{pl_compAB}
\alias{pl_compAB}
\title{pl_compAB}
\usage{
pl_compAB(agr_df, which = "D", cst_df = NULL, pl_labs = NULL)
}
\arguments{
\item{agr_df}{Dataframe of timestamped comparisons between two coders}

\item{which}{Determines whether to plot agreements ("A"), or disagreements ("D")}

\item{cst_df}{Optional dataframe of work-focus codestamps.}

\item{pl_labs}{Optional ggplot plotting lables}
}
\value{
A ggplot plotting object showing the comparison between two code datasets.
}
\description{
This function creates a standard session plot object, markes out the intervals of coder disagreements using top-to-bottom grey bars, and optinally overlays a set of codestamps.
}
