% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{wkf_build_dsdesc}
\alias{wkf_build_dsdesc}
\title{wkf_build_dsdesc}
\usage{
wkf_build_dsdesc(dslist, change = NULL)
}
\arguments{
\item{dslist}{A dataset descriptor list with either the name element or a set of initial dataset identifiers present.}

\item{change}{A list specifying how dataset identifiers should be replaced.}
}
\value{
A dataset descriptor list. The first element (name) will be the dataset name createc by applying any change elements to the description in `dslist`. Elements 2:7 (sid, type, method, coder, version, and stampe) will be the component identifiers of that name. Any identifiers unspecied by `dslist$name` and `change` will be "".
}
\description{
Builds a dataset descriptor list from an existing dataset descriptor and specified dataset identifiers.
}
\examples{
dsid <- list("", "res2C", "focus", "video", "mdg", "20180101", "cstamp")
names(dsid) <- c("name", "sid", "type", "method", "coder", "version", "stamp")
change <- list(coder = "slg", version = "20180102")
# Build a name from identfier list with changes
wkf_build_dsdesc(dsid, change)
#
# Build a name from identifiers list without change
dsid$name <- ""
wkf_build_dsdesc(dsid, change = NULL)
#
# Build descriptor by swapping parts of a ds name
dsid$name <- "res1A_focus_hand_slg_20180101_tsample"
dsid[2:7] <- ""
wkf_build_dsdesc(dsid, list(version = "20171228"))
\dontrun{
# Poor specification - embedded identifier is missing
dsid$name <- "res2C_focus_hand"
dsid[2:7] <- ""
wkf_build_dsdesc(dsid, list(version = "20171228"))}
# Correct
dsid$name <- "res2C_focus_hand"
dsid[2:7] <- ""
wkf_build_dsdesc(dsid, list(coder = "slg", version = "20171228", stamp = ""))
}
