% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wkFocus-tsample.R
\name{wkf_comp_codings}
\alias{wkf_comp_codings}
\title{wkf_comp_codings}
\usage{
wkf_comp_codings(A, B)
}
\arguments{
\item{{A, }{B} Work focus codes. NA indicates no code.}
}
\value{
A match indicator. One of "A" (agree), "D" (disagree), or "N" (nothing
}
\description{
This function implements part of the decision algorithm being used to count how many times two coders coding for up to two focuses agree.
}
\examples{
wkf_comp_codings("Rd", "Rd")  # returns "A"
wkf_comp_codings("Rd", "Do")  # returns "D"
wkf_comp_codings("Rd",  NA)   # returns "D"
wkf_comp_codings( NA,   NA)   # returns "N"
}
