% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compAB.R
\name{wkf_compAB_tsample}
\alias{wkf_compAB_tsample}
\title{wkf_compAB_tsample}
\usage{
wkf_compAB_tsample(tsAB)
}
\arguments{
\item{tsAB}{A list containing timesamples for two coders, A and B.}
}
\value{
A list giving the time and the two agreement decisions at that time.
}
\description{
Compares coding decisions made by two coders (A, B) at a given time.
}
\examples{
# One agreement
tsAB1 <- list(A.t = t1, A.focus1 = "Rd", A.focus2 = NA_character_,
              B.t = t1, B.focus1 = "Rd", B.focus2 = NA_character_)
wkf_compAB_tsample(tsAB1)
# Two agreements - order doesn't matter
tsAB2 <- list(A.t = t1, A.focus1 = "Rd", A.focus2 = "Do",
              B.t = t1, B.focus1 = "Do", B.focus2 = "Rd")
wkf_compAB_tsample(tsAB2)
# One agreement, one disagreement
tsAB3 <- list(A.t = t1, A.focus1 = "Rd", A.focus2 = "Do",
              B.t = t1, B.focus1 = "Rd", B.focus2 = NA_character_)
wkf_compAB_tsample(tsAB3)
# One agreement, one disagreement
list(A.t = t1, A.focus1 = "Rd", A.focus2 = "Do",
     B.t = t1, B.focus1 = "Ds", B.focus2 = "Rd") \%>\%
  wkf_compAB_tsample()
# Two disagreements
list(A.t = t1, A.focus1 = "Rd", A.focus2 = "Do",
     B.t = t1, B.focus1 = "Rp", B.focus2 = "Ds") \%>\%
  wkf_compAB_tsample()
\dontrun{
# Coder B has duplicte code - error
tsAB_err <- list(A.t = t1, A.focus1 = "Rd", A.focus2 = NA_character_,
                 B.t = t1, B.focus1 = "Rd", B.focus2 = "Rd")
wkf_compAB_tsample(tsAB_err)}
# Can use `map()` to produce a data frame of decisions
tsAB1_3 <- as.data.frame(rbind(tsAB1, tsAB2, tsAB3))
tsAB1_3 \%>\% purrr::transpose() \%>\%
  purrr::map_df(wkf_compAB_tsample)
}
