% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wkf-utils.R
\name{wkf_config}
\alias{wkf_config}
\title{wkf_config}
\usage{
wkf_config()
}
\value{
A list of named parameters configuring this project
}
\description{
Provides the project configuration file establishing global parameters such as
plotting scales, color codes for factor levels, etc.
}
