% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{wkf_convert_tcode}
\alias{wkf_convert_tcode}
\title{wkf_convert_tcode}
\usage{
wkf_convert_tcode(tcode, fr, origin)
}
\arguments{
\item{tcode}{timecode <chr> of form "hh:mm:ss:ff".}

\item{fr}{the frame rate the for the video timecode}

\item{origin}{a <chr> giving a starting point for the timestamps}
}
\value{
dataframe with four variables: three <chr> giving the original timecode, time only, and the frames only, and one <POSIXct> giving the time from the origin.
}
\description{
Converts timecodes into a number of seconds.
}
\examples{
beg <- "2016-08-16 00:00:00 UTC"
timecode <- "00:12:25:12"
str(wkf_convert_tcode(timecode, 25, beg))
# Round up based on framecount
timecode <- "00:12:25:20"
str(wkf_convert_tcode(timecode, 25, beg))
}
