% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wkFocus-utils.R
\name{wkf_parse_dspath}
\alias{wkf_parse_dspath}
\title{wkf_parse_dspath}
\usage{
wkf_parse_dspath(path)
}
\arguments{
\item{path}{Path to a file holding a work focus dataset, or the name of a dataset. The dataset name should be of the form sid-type-method-coder-version-stamp. However, training identifiers can be missing.}
}
\value{
A list with the full dataset name (`name`), the session ID (`sid`), the code type (`type`),the method (`method`), the coder (`coder`), the version (`version`), and the type of stamp (`stamp`). Missing trailing identifiers will return as "".
}
\description{
Extracts the file name in `path`, excluding any extention, and parses it for component identifiers of a work-focus coding dataset. The identifiers are located soley by their position in the file name.
}
\examples{
path <- "~/data/res2C-focus-hand-mdg-20180101-cstamp.Rda"
wkf_parse_dspath(path)
# Does not need directory specifier
path <- "res1A-focus-video.xlsx"
wkf_parse_dspath(path)
# Works for bare dsname wo/extension
path <- "res1A-focus-hand-slg-20180101-tsample"
wkf_parse_dspath(path)
\dontrun{
# Can't have missinng ds IDs in the middle
path <- "video-focus-video--20180101.edl"
wkf_parse_dspath(path)}
}
