% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{wkf_parse_sid}
\alias{wkf_parse_sid}
\title{wkf_parse_sid}
\usage{
wkf_parse_sid(sid)
}
\arguments{
\item{sid}{A single character string showing the session ID, "res1A", etc.}
}
\value{
A list of descriptor IDs characterizing the session.
}
\description{
Breaks a session ID into component identifiers for the session and returns a list of identifiers that can be used to build a dataset descriptor. The results depends of which phase the session is in.
}
\examples{
wkf_parse_sid("res1A")
}
