% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{wkf_pl_compAB}
\alias{wkf_pl_compAB}
\title{wkf_pl_compAB}
\usage{
wkf_pl_compAB(agr_df, which = "D", cst_df = NULL, pl_labs = NULL)
}
\arguments{
\item{agr_df}{Dataframe of timestamped comparisons between two coders}

\item{which}{Determines whether to plot agreements ("A"), or disagreements ("D")}

\item{cst_df}{Optional dataframe of work-focus codestamps.}

\item{pl_labs}{Optional ggplot plotting lables}
}
\value{
A ggplot plotting object showing the comparison between two code datasets.
}
\description{
Plots a dataframe of comparisons between two coders (A, B) and optinally overlays a set of focus codes.
}
