% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{wkf_pl_comps}
\alias{wkf_pl_comps}
\title{wkf_pl_comps}
\usage{
wkf_pl_comps(ds_agrAB, to_plot = "D", ds_stack = NULL, pl_labs = NULL)
}
\arguments{
\item{ds_agrAB}{A single dataframe of timestamped comparisons between two coders}

\item{to_plot}{Determines whether to plot agreements ("A"), or disagreements ("D")}

\item{ds_stack}{A single dataframe of stacked codestamps to overlay on the plot.}

\item{pl_labs}{Optional ggplot plotting lables}
}
\value{
A ggplot plotting object showing the comparison between two code datasets.
}
\description{
Plots a set of comparisons between two coders (A, B) and optinally overlays a set of focus codes.
}
