% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{wkf_pl_cstamps}
\alias{wkf_pl_cstamps}
\title{wkf_pl_cstamps}
\usage{
wkf_pl_cstamps(ds_stack, pl_labs = NULL)
}
\arguments{
\item{ds_stack}{A single dataframe of stacked cstamps}

\item{pl_labs}{An optional list of ggplot2 plotting labels}
}
\value{
A ggplot object showing how the work-focus codes in `ds_stack` are distributed over session time.
}
\description{
Will graph segments for all codes in the database. Uses a generic title. Add subtitle, y-axis label, etc. through `pl_labs` if wanted.
}
