% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{wkf_pl_focus}
\alias{wkf_pl_focus}
\title{wkf_pl_focus}
\usage{
wkf_pl_focus(cst_df, pl_labs = NULL)
}
\arguments{
\item{cst_df}{A standard codestamp data frame}

\item{pl_labs}{An optional list of ggplot2 plotting labels}
}
\value{
A ggplot object showing how the work-focus codes in `cst_df` are distributed over session time.
}
\description{
Will graph segments for all codes in the database. Uses a generic title. Add subtitle, y-axis label, etc. through `pl_labs` if wanted.
}
