% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsample.R
\name{wkf_tsample}
\alias{wkf_tsample}
\title{wkf_tsample}
\usage{
wkf_tsample(ds_cst, dt = 1, bt = NULL, et = NULL, verbose = FALSE,
  warn = TRUE)
}
\arguments{
\item{ds_cst}{A codestamp dataset ($type = "cstamp").}

\item{dt}{Sampling timestep. Defaults to 1 second.}

\item{bt}{POSIXct object giving the time to begin sampling. If NULL, sampling will begin at the earliest time in the dataset.}

\item{et}{POSIXct object given the time to end sampling. If NULL, sampling will end at the latest time in the dataset.}

\item{verbose}{If TRUE, codes will be printed at each timestep and a summary of missing codes will be given.}

\item{warn}{Logical. If true (default), warnings will be issued when an NA code}
}
\value{
A time-sample dataset ($type = "tsample").
}
\description{
Samples a single cstamp dataset between two time stamps using a specified time step.
}
\examples{
smpl_codestamps <- res1A_focus_hand_cstamp[[1]]
time_samples <- wkf_tsample(smpl_codestamps)
str(time_samples)
}
