% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsamples.R
\name{wkf_tsample_cstamps}
\alias{wkf_tsample_cstamps}
\title{wkf_tsample_cstamps}
\usage{
wkf_tsample_cstamps(ds, dt = 1, bt = NULL, et = NULL, verbose = FALSE,
  warn = TRUE)
}
\arguments{
\item{ds}{A list with ds$name = dataset name and ds$data = a data frame of codestamps.}

\item{dt}{Sampling timestep. Defaults to 1 second.}

\item{bt}{POSIXct object giving the time to begin sampling. If NULL, sampling will begin at the earliest time in the dataset.}

\item{et}{POSIXct object given the time to end sampling. If NULL, sampling will end at the latest time in the dataset.}

\item{verbose}{If TRUE, codes will be printed at each timestep and a summary of missing codes will be given.}
}
\value{
A data frame of time-samples with the "desc" attribute set to the descriptor of the dataset sampled.
}
\description{
wkf_tsample_cstamps
}
\examples{
smpl_codestamps <- res1A_focus_hand_cstamp[[1]]
time_samples <- wkf_tsample_cstamps(smpl_codestamps)
str(time_samples)
}
