% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wkFocus-tsample.R
\name{wkf_tsamples_agree}
\alias{wkf_tsamples_agree}
\title{wkf_tsamples_agree}
\usage{
wkf_tsamples_agree(ts1, ts2)
}
\arguments{
\item{codeset}{A list of length 5, giving the time and two pairs of coder}
}
\value{
A list giving the time and the two agreement decisions at that time.
}
\description{
This function implements part of the decision algorithm  being
}
