% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wkFocus-package.R
\docType{package}
\name{wkFocus}
\alias{wkFocus}
\alias{wkFocus-package}
\title{wkFocus: A package for analyzing the work focus for the "research phase" of my dissertation study.}
\description{
The wkFocus package comprises three parts: the focus-code datasets, the functions for working with these databases, and several vignettes exploring the databases.
}
\section{wkFocus datasets}{

The wkFocus datasets ...
}

\section{wkFocus functions}{

The wkFocus functions ...
}

\section{wkFocus vignettes}{

The wkFocus vignettes ...
}

