% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wkf-utils.R
\name{wkf_parse_dspath}
\alias{wkf_parse_dspath}
\title{wkf_parse_dspath}
\usage{
wkf_parse_dspath(path = "")
}
\arguments{
\item{path}{Path to a file holding a work focus dataset, or the name of a dataset. The dataset name should be of the form sid-type-method-coder-version-stamp. Trailing identifiers can be missing.}
}
\value{
A list with the full dataset name (`name`), the session ID (`sid`), the code type (`type`),the method (`method`), the coder (`coder`), the version (`version`), and the type of stamp (`stamp`). Missing trailing identifiers will return as "". An empty file path returns a descriptor with empty identifiers.
}
\description{
Extracts the file name in `path`, excluding any extention, and parses it for component identifiers of a work-focus coding dataset. The identifiers are located soley by their position in the file name.
}
\examples{
path <- "~/data/res2C_focus_hand_mdg_20180101_cstamp.Rda"
wkf_parse_dspath(path)
# Does not need directory specifier
path <- "res1A_focus_video.xlsx"
wkf_parse_dspath(path)
# Works for bare dsname wo/extension
path <- "res1A_focus_hand_slg_20180101_tsample"
wkf_parse_dspath(path)
\dontrun{
# Can't have missinng ds IDs in the middle
path <- "video_focus_video__20180101.edl"
wkf_parse_dspath(path)}
}
